gun: merge Gun_GetNewTarget and Gun_ChangeTarget
